Step 3 # Bootstrap the application using standalone APIs - Deleted app module.ts and added them to main.ts
